% File src/library/corrsieve/man/QmatrixFilt.Rd
\name{QmatrixFilt}
\alias{QmatrixFilt-class}
\title{QmatrixFilt}
\description{The S4 class QmatrixFilt lists for Q matrix correlation output}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("QmatrixFilt", ...)}.
}
\section{Slots}{
  \describe{
	\item{rowncol}{A list listing filtered Q matrix correlations by the rows-and-columns method}
	\item{avmaxcorr}{A table listing filtered Q matrix correlations by the rows-and-columns method}
	\item{rawcorr}{A list listing raw Q matrix correlations}
}}
\author{Michael G. Campana <mgc32@cam.ac.uk>}
\seealso{
	\code{\link{QmatrixFilt}}
}
\keyword{Q matrix}
\keyword{Qmatrix}
\keyword{STRUCTURE}
\keyword{Structure}
\keyword{structure}
