% File src/library/corrsieve/man/summarise.Fst.Rd
\name{summarise.Fst}
\alias{summarise.Fst}
\alias{summarize.Fst}
\title{Summarise.Fst}
\description{Summarises Fst from structure output read by \code{read.struct}.}
\usage{summarise.Fst(input)}
\arguments{
	\item{input}{a table containing lnPD Fst generated by \code{read.struct}}
}
\value{Returns a table listing K values and summarised Fst statistics}
\author{Michael G. Campana <mgc32@cam.ac.uk>}
\seealso{
	\code{\link{read.struct}}
	\code{link{calc.delta}}
}
\keyword{summarise}
\keyword{summarize}
\keyword{Fst}
\keyword{fst}
\keyword{STRUCTURE}
\keyword{Structure}
\keyword{structure}
