% File src/library/corrsieve/man/summarise.lnPD.Rd
\name{summarise.lnPD}
\alias{summarise.lnPD}
\alias{summarize.lnPD}
\title{Summarise.lnPD}
\description{Summarises lnP(D) from structure output read by \code{read.struct}.}
\usage{summarise.lnPD(input)}
\arguments{
	\item{input}{a table containing lnPD data generated by \code{read.struct}}
}
\value{Returns a table listing K values and summarised lnPD statistics}
\author{Michael G. Campana <mcampana63@gmail.com>}
\seealso{
	\code{\link{read.struct}}
	\code{\link{calc.delta}}
}
\keyword{summarise}
\keyword{summarize}
\keyword{lnPD}
\keyword{lnpd}
\keyword{lnP(D)}
\keyword{STRUCTURE}
\keyword{Structure}
\keyword{structure}
