% File src/library/corrsieve/man/summarise.Fst.Rd
\name{summarise.Fst}
\alias{summarise.Fst}
\alias{summarize.Fst}
\title{Summarise.Fst}
\description{Summarises Fst from structure output read by \code{read.struct}.}
\usage{summarise.Fst(input, stdevopt = 1)}
\arguments{
	\item{input}{a table containing lnPD Fst generated by \code{read.struct}}
	\item{stdevopt}{Chooses the optimisation procedure for the Fst summaries. 
		1: no optimisation, 2: order the clusters by value, 3: order the clusters by correlation coefficients}
}
\value{Returns a table listing K values and summarised Fst statistics}
\author{Michael G. Campana <mcampana63@gmail.com>}
\seealso{
	\code{\link{read.struct}}
	\code{\link{calc.delta}}
}
\keyword{summarise}
\keyword{summarize}
\keyword{Fst}
\keyword{fst}
\keyword{STRUCTURE}
\keyword{Structure}
\keyword{structure}
